requirements.txt
google-generativeai>=0.3.0
google-genai>=0.2.0
python-dotenv>=1.0.0
googlemaps>=4.10.0  # For Google Places API (optional - currently using mock data)

# Calendar Integration Dependencies
google-api-python-client>=2.100.0  # Calendar & Gmail APIs
google-auth-httplib2>=0.1.1       # HTTP transport
google-auth-oauthlib>=1.1.0        # OAuth flow
schedule>=1.2.0                    # Task scheduling

